requirements.txt
streamlit
easyocr
torch
torchvision
opencv-python-headless
numpy
Pillow
openpyxl
ultralytics
